Fix microphone not working after second flip on Chrome
Previously, the microphone input view only worked the first time it was
accessed. Starting with the second time, it would stay stuck on the same
pitch. It is unclear why, but I needed to recreate the analyser node
each time when flipping to the microphone input view for it to work.