feat: improve UI and rewrite in Lit (#22)
This is a complete rewrite of the app using web components (Lit) and other
modern web tooling such as Vite. From a user perspective, the main change is the
revamped UI, which is much more modern and removes some of the pain points with
the previous UI, such as starting in a non-functional state by default (not
listening for a pitch or playing anything).